Update sha1.c
Changing int to uint8_t.

+++ Arduino_Code/sha1/sha1.c
@@ -19,7 +19,7 @@
 ssize_t sha1_hasher_write(sha1_hasher_t hasher, const void * buf, size_t count)
 {
 	size_t written = 0;
-	int chk_result;
+	uint8_t chk_result;
 	char c;
 	while(written < count)
 	{
